% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_funs.R
\name{docker_check_install}
\alias{docker_check_install}
\title{Check if Docker is installed}
\usage{
docker_check_install()
}
\value{
Logical
}
\description{
Check if Docker is installed
}
\examples{
library(RFsel)
docker_check_install()
}
