% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_funs.R
\name{docker_ps}
\alias{docker_ps}
\title{List Docker Containers}
\usage{
docker_ps(type = c("running", "exited", "all"))
}
\arguments{
\item{type}{running / exited / all}
}
\value{
A Dataframe
}
\description{
List Docker Containers
}
\examples{
library(RFsel)
docker_ps("running")
docker_ps("exited")
docker_ps("all")
}
