% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_funs.R
\name{docker_remove_container}
\alias{docker_remove_container}
\title{Stop Remove Containers}
\usage{
docker_remove_container(ids)
}
\arguments{
\item{ids}{A Character Vector of Conteiner IDs or Names}
}
\value{
Success Message
}
\description{
Stop Remove Containers
}
