% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_funs.R
\name{docker_restart_container}
\alias{docker_restart_container}
\title{Stop Restart Containers}
\usage{
docker_restart_container(ids)
}
\arguments{
\item{ids}{A Character Vector of Conteiner IDs or Names}
}
\value{
Success Message
}
\description{
Stop Restart Containers
}
