% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_funs.R
\name{docker_stop_container}
\alias{docker_stop_container}
\title{Stop Docker Containers}
\usage{
docker_stop_container(ids)
}
\arguments{
\item{ids}{A Character Vector of Conteiner IDs or Names}
}
\value{
Success Message
}
\description{
Stop Docker Containers
}
