% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_funs.R
\name{download_js}
\alias{download_js}
\title{Helper: Download a file via Javascript Injection}
\usage{
download_js(.rsc, .url, .js, .dir_map, .dir_doc, .wait = 10)
}
\arguments{
\item{.rsc}{A Remote Selenium Connection (rsc): Use sel_get_driver()}

\item{.url}{A Url}

\item{.js}{A Javascript to be injected}

\item{.dir_map}{Mapped directory of the Docker Container (see: sel_docker_start())}

\item{.dir_doc}{Directory to save files}

\item{.wait}{Wait time before error}
}
\value{
A downloaded file and a character vector (path of the file is returned)
}
\description{
Helper: Download a file via Javascript Injection
}
