% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_funs.R
\name{extract_doc_id}
\alias{extract_doc_id}
\title{Helper: Extract Doc ID}
\usage{
extract_doc_id(.path)
}
\arguments{
\item{.path}{path to the file}
}
\value{
A character (Doc ID)
}
\description{
Helper: Extract Doc ID
}
