% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{firefox_ecaps_direct_download}
\alias{firefox_ecaps_direct_download}
\title{Extracapabilities for Firefox Browser (Direct Download)}
\format{
List with 7 elements
\describe{
  \item{browser.download.dir}{TBD}
  \item{browser.download.folderList}{TBD}
  \item{browser.download.manager.showWhenStarting}{TBD}
  \item{browser.helperApps.neverAsk.saveToDisk}{TBD}
  \item{browser.helperApps.neverAsk.openFile}{TBD}
  \item{browser.helperApps.alwaysAsk.force}{TBD}
  \item{pdfjs.disabled}{TBD}
  ...
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
firefox_ecaps_direct_download
}
\description{
A List specifying ecaps for Firefox
}
\keyword{datasets}
