% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_funs.R
\name{lft2}
\alias{lft2}
\title{List files in a dataframe with id column (Updated Function)}
\usage{
lft2(dirs, reg = "*", id = "doc_id", rec = FALSE, info = FALSE)
}
\arguments{
\item{dirs}{Paths to the directories}

\item{reg}{RegEx to find files (defaults to '*' all files)}

\item{id}{Column name containing the file name (e.g. doc_id)}

\item{rec}{Should the directories be searched recursively?}

\item{info}{Additional Infos (base::)}
}
\value{
A dataframe with file paths
}
\description{
A helper function that is build around list.files (base)
}
