% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_funs.R
\name{map_download_js}
\alias{map_download_js}
\title{Download a files via Javascript Injection}
\usage{
map_download_js(.rsc, .urls, .jss, .dir_map, .dir_doc, .wait = 10)
}
\arguments{
\item{.rsc}{A Remote Selenium Connection (rsc): Use sel_get_driver()}

\item{.urls}{A vector of URls}

\item{.jss}{A vector of Javascripts to be injected}

\item{.dir_map}{Mapped directory of the Docker Container (see: sel_docker_start())}

\item{.dir_doc}{Directory to save files}

\item{.wait}{Wait time before error}
}
\value{
A Dataframe
}
\description{
Download a files via Javascript Injection
}
