% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pia_funs.R
\name{pia_connect}
\alias{pia_connect}
\title{PIA connect}
\usage{
pia_connect(piactl = NULL)
}
\arguments{
\item{piactl}{default: C:/Program Files/Private Internet Access/piactl.exe}
}
\value{

}
\description{
PIA connect
}
