% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pia_funs.R
\name{pia_disconnect}
\alias{pia_disconnect}
\title{PIA disconnect}
\usage{
pia_disconnect(piactl = NULL)
}
\arguments{
\item{piactl}{default: C:/Program Files/Private Internet Access/piactl.exe}
}
\value{
Nothing, disonnects PIA
}
\description{
PIA disconnect
}
