% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pia_funs.R
\name{pia_get_connection_state}
\alias{pia_get_connection_state}
\title{PIA get Connection State}
\usage{
pia_get_connection_state(piactl = NULL)
}
\arguments{
\item{piactl}{default: C:/Program Files/Private Internet Access/piactl.exe}
}
\value{

}
\description{
PIA get Connection State
}
