% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pia_funs.R
\name{pia_get_regions}
\alias{pia_get_regions}
\title{PIA get Regions}
\usage{
pia_get_regions(piactl = NULL)
}
\arguments{
\item{piactl}{default: C:/Program Files/Private Internet Access/piactl.exe}
}
\value{
A character vector
}
\description{
PIA get Regions
}
