% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pia_funs.R
\name{pia_get_vpn_ip}
\alias{pia_get_vpn_ip}
\title{PIA get VPN IP}
\usage{
pia_get_vpn_ip(piactl = NULL)
}
\arguments{
\item{piactl}{default: C:/Program Files/Private Internet Access/piactl.exe}
}
\value{
A caracter, PIA IP Address
}
\description{
PIA get VPN IP
}
