% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selenium_funs.R
\name{sel_docker_start}
\alias{sel_docker_start}
\title{Start Docker Selenium Container}
\usage{
sel_docker_start(
  type = c("debug", "static"),
  browser = c("firefox", "chrome"),
  version = "latest",
  port_sel = 4445,
  port_vcn = 5901,
  dir_map = NULL
)
}
\arguments{
\item{type}{Debug/Static}

\item{browser}{Firefox/Chrome}

\item{version}{Browser Version}

\item{port_sel}{Selenium Port (default: 4445)}

\item{port_vcn}{VCN Port (default: 5901)}

\item{dir_map}{Mapping Directory}
}
\value{
A List
}
\description{
Start Docker Selenium Container
}
