% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pia_funs.R
\name{pia_set_region}
\alias{pia_set_region}
\title{PIA set Region}
\usage{
pia_set_region(piactl = NULL, region)
}
\arguments{
\item{piactl}{default: C:/Program Files/Private Internet Access/piactl.exe}

\item{region}{A Region (see pia_get_regions())}
}
\value{
-
}
\description{
PIA set Region
}
